+++ changelog.md
@@ -1,7 +1,12 @@
 # Changelog
 
-## 3.4.2 (2020-05-06)
+## 3.4.3 (2020-05-06)
 * Small styling adjustments for "Basic theme" and "Recras Blue" themes
+
+## 3.4.2 (2020-04-24)
+* Fix contact form placeholders generating invalid HTML
+* Email/Telephone fields in contact forms didn't get proper field type - fixed
+* Improve accessibility and styling of required field labels
 
 ## 3.4.1 (2020-04-16)
 * Updated German translations, courtesy of Wiljon Bolten